MadeTheGame
I reworked an old project and replaced the empty game with this project. I then completely reshaped the project by changing tons of code, deleting and replacing objects and sprites, and such. This will game should work just fine but there are a couple things to do and plenty we can add to it.